se agrego validacion al emliminar un cliente

+++ src/main/java/com/example/entregafinal/service/CustomerService.java
@@ -2,7 +2,6 @@ package com.example.entregafinal.service;
 
 import com.example.entregafinal.model.Customer;
 import com.example.entregafinal.repository.CustomerRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,6 +9,7 @@ import java.util.Optional;
 
 @Service
 public class CustomerService {
+
     @Autowired
     private CustomerRepository customerRepository;
 
@@ -18,7 +18,11 @@ public class CustomerService {
     }
 
     public void eliminarCustomer(int id) {
-        customerRepository.deleteById(id);
+        Optional<Customer> customer = customerRepository.findById(id);
+        if (customer.isPresent()) {
+            customerRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Cliente con ID " + id + " no encontrado");
+        }
     }
-
 }
